+++ jetbrains-core/resources/META-INF/plugin.xml
@@ -90,6 +90,8 @@ with what features/services are supported.
                   topic="software.aws.toolkits.jetbrains.core.credentials.ConnectionSettingsStateChangeNotifier"/>
         <listener class="software.aws.toolkits.jetbrains.services.dynamic.DynamicResourceStateChangedNotificationHandler"
                   topic="software.aws.toolkits.jetbrains.services.dynamic.DynamicResourceStateMutationHandler"/>
+        <listener class="software.aws.toolkits.jetbrains.services.dynamic.DynamicResourcesRefreshVirtualFileHandler"
+                  topic="software.aws.toolkits.jetbrains.services.dynamic.DynamicResourceStateMutationHandler"/>
 
     </projectListeners>
     <extensionPoints>
