+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/dynamic/DynamicResourceStateChangedNotificationHandler.kt
@@ -4,15 +4,11 @@
 package software.aws.toolkits.jetbrains.services.dynamic
 
 import com.intellij.openapi.application.runInEdt
-import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.project.Project
-import com.intellij.psi.PsiManager
-import software.amazon.awssdk.services.cloudcontrol.model.Operation
 import software.amazon.awssdk.services.cloudcontrol.model.OperationStatus
 import software.aws.toolkits.jetbrains.core.AwsResourceCache
 import software.aws.toolkits.jetbrains.core.explorer.ExplorerToolWindow
 import software.aws.toolkits.jetbrains.services.dynamic.DynamicResourceTelemetryResources.addOperationToTelemetry
-import software.aws.toolkits.jetbrains.services.dynamic.explorer.DynamicResourceNode
 import software.aws.toolkits.jetbrains.utils.notifyError
 import software.aws.toolkits.jetbrains.utils.notifyInfo
 import software.aws.toolkits.resources.message
@@ -25,9 +21,6 @@ class DynamicResourceStateChangedNotificationHandler(private val project: Projec
     private val refreshRequired = AtomicBoolean(false)
     override fun mutationStatusChanged(state: ResourceMutationState) {
         if (state.status == OperationStatus.SUCCESS) {
-            if (state.operation == Operation.UPDATE) {
-                refreshViewEditableDynamicResourceVirtualFile(state)
-            }
             notifyInfo(
                 message(
                     "dynamic_resources.operation_status_notification_title",
@@ -79,19 +72,6 @@ class DynamicResourceStateChangedNotificationHandler(private val project: Projec
         }
         AwsResourceCache.getInstance().clear(CloudControlApiResources.listResources(state.resourceType), state.connectionSettings)
         refreshRequired.set(true)
-    }
-
-    private fun refreshViewEditableDynamicResourceVirtualFile(state: ResourceMutationState) {
-        val file = state.resourceIdentifier?.let { DynamicResourceIdentifier(state.connectionSettings, state.resourceType, it) }?.let {
-            CloudControlApiResourcesUtils.getResourceFile(project,
-                it
-            )
-        } as? ViewEditableDynamicResourceVirtualFile ?: return
-        file.isWritable = true
-        runInEdt {
-            val psiFile = PsiManager.getInstance(project).findFile(file)
-            psiFile?.text?.let { file.setContent(this, it, true) }
-        }
     }
 
     private fun displayErrorMessage(state: ResourceMutationState, errorMessage: String) {
